+++ src/pages/Character/components/ItemGrid.styles.ts
