+++ src/pages/Character/components/ItemModal.styles.ts
