+++ src/pages/Character/components/Tab.tsx
